django/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-dsfr==3.0.0
django-revproxy==0.13.0
djangorestframework==3.16.0
djangorestframework-api-key==3.1.0
factory_boy==3.3.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
responses==0.25.7
ruff==0.12.2
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.9.0